move CValidationDebugger to its own header outside of consensus
